+++ SwiftCharts/Views/CubicLinePathGenerator.swift
@@ -10,23 +10,30 @@ import UIKit
 
 public class CubicLinePathGenerator: ChartLinesViewPathGenerator {
     
-    public init() {}
+    private let tension1: CGFloat
+    private let tension2: CGFloat
     
-    // src: http://stackoverflow.com/a/29876400/930450
+    /**
+    - parameter tension1: p1 tension, where 0 is straight line. A value higher than 0.3 is not recommended.
+    - parameter tension2: p2 tension, where 0 is straight line. A value higher than 0.3 is not recommended.
+    */
+    public init(tension1: CGFloat, tension2: CGFloat) {
+        self.tension1 = tension1
+        self.tension2 = tension2
+    }
+    
+    // src: http://stackoverflow.com/a/29876400/930450 (modified)
     public func generatePath(points points: [CGPoint], lineWidth: CGFloat) -> UIBezierPath {
-        var cp1: CGPoint
-        var cp2: CGPoint
         
         let path = UIBezierPath()
         var p0: CGPoint
         var p1: CGPoint
         var p2: CGPoint
         var p3: CGPoint
-        var tensionBezier1: CGFloat = 0.3
-        var tensionBezier2: CGFloat = 0.3
+        var tensionBezier1: CGFloat
+        var tensionBezier2: CGFloat
         
         var previousPoint1: CGPoint = CGPointZero
-//        var previousPoint2: CGPoint
         
         path.moveToPoint(points.first!)
         
@@ -34,9 +41,8 @@ public class CubicLinePathGenerator: ChartLinesViewPathGenerator {
             p1 = points[i]
             p2 = points[i + 1]
             
-            let maxTension: CGFloat = 1 / 3
-            tensionBezier1 = maxTension
-            tensionBezier2 = maxTension
+            tensionBezier1 = self.tension1
+            tensionBezier2 = self.tension2
             
             if i > 0 {  // Exception for first line because there is no previous point
                 p0 = previousPoint1
@@ -59,28 +65,13 @@ public class CubicLinePathGenerator: ChartLinesViewPathGenerator {
                 p3 = p2
                 tensionBezier2 = 0
             }
+
+            let controlPoint1 = CGPointMake(p1.x + (p2.x - p1.x) / 3, p1.y - (p1.y - p2.y) / 3 - (p0.y - p1.y) * tensionBezier1)
+            let controlPoint2 = CGPointMake(p1.x + 2 * (p2.x - p1.x) / 3, (p1.y - 2 * (p1.y - p2.y) / 3) + (p2.y - p3.y) * tensionBezier2)
             
-            // The tension should never exceed 0.3
-            if tensionBezier1 > maxTension {
-                tensionBezier1 = maxTension
-            }
-            if tensionBezier1 > maxTension {
-                tensionBezier2 = maxTension
-            }
-            
-            // First control point
-            cp1 = CGPointMake(p1.x + (p2.x - p1.x)/3,
-                p1.y - (p1.y - p2.y)/3 - (p0.y - p1.y)*tensionBezier1)
-            
-            // Second control point
-            cp2 = CGPointMake(p1.x + 2*(p2.x - p1.x)/3,
-                (p1.y - 2*(p1.y - p2.y)/3) + (p2.y - p3.y)*tensionBezier2)
-            
-            
-            path.addCurveToPoint(p2, controlPoint1: cp1, controlPoint2: cp2)
+            path.addCurveToPoint(p2, controlPoint1: controlPoint1, controlPoint2: controlPoint2)
             
             previousPoint1 = p1;
-//            previousPoint2 = p2;
         }
         
         return path