+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 - Fix tracker showing NaN on constant coordinate
+- Pass tension values to cubic line to be able to modify curviness
 
 ## [0.3] - 2015.09.28
 - Merge swift2.0 in master